pyproject.toml
[project]
name = "credit-scout"
version = "0.1.0"
description = "AI agent for movie credit detection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "rich>=13.9.4",
    "loguru>=0.7.2",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "google-genai>=1.16.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/credit_scout/__version__.py"

[project.scripts]
credit-scout = "credit_scout.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/credit-scout"]

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default
lint.select = ["E", "F", "I"]
# Allow unused variables when underscore-prefixed
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
lint.ignore = ["E272"]
# Python version compatibility
target-version = "py312"
# Line length matches black's default
line-length = 120

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]
